unify rust apps as subcommands
