--- performance/comparison/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "comparison"
-version = "0.1.0"
-edition = "2018"
-
-[dependencies]
-itertools = "0.10.1"
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-thiserror = "1.0.26"
